[MODEL-23520] -- refactor
