+++ src/datarobot_genai/drmcp/core/feature_flags.py
@@ -13,8 +13,10 @@
 # limitations under the License.
 from async_lru import alru_cache
 
+from datarobot_genai.drmcpbase.datarobot_services.feature_flags import (
+    is_mcp_tools_gallery_support_enabled,
+)
 from datarobot_genai.drtools.core.credentials import get_credentials
-from datarobot_genai.drtools.core.feature_flags import is_mcp_tools_gallery_support_enabled
 
 
 class FeatureFlag:
